requirements.txt
aioredis==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.2.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.112.0
fastapi-cache2==0.2.2
greenlet==3.0.3
h11==0.14.0
hiredis==3.0.0
idna==3.7
Mako==1.3.5
MarkupSafe==2.1.5
passlib==1.7.4
pendulum==3.0.0
psycopg==3.2.1
pydantic==2.8.2
pydantic[email]
pydantic_core==2.20.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
redis==4.6.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
time-machine==2.15.0
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6